+++ lib/mdx.ml
@@ -26,6 +26,7 @@ module Migrate_ast = Migrate_ast
 module Compat = Compat
 module Util = Util
 module Prelude = Prelude
+module Syntax = Syntax
 
 type line =
   | Section of (int * string)